pyproject.toml
[project]
name = "q3-assignments"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "nest-asyncio>=1.6.0",
    "openai-agents>=0.2.9",
]
